+++ src/main/java/com/isst/mystay/model/Empleado.java
@@ -10,31 +10,65 @@ public class Empleado {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
 
     private String nombre;
     private String funcion;
     private String telefono;
     private String correo;
-    private long hotelId;
+    private Integer hotelId;
 
     // Constructor por defecto
     public Empleado() {
     }
 
     // Getters
-    public long getId() { return id; }
-    public String getNombre() { return nombre; }
-    public String getFuncion() { return funcion; }
-    public String getTelefono() { return telefono; }
-    public String getCorreo() { return correo; }
-    public long getHotelId() { return hotelId; }
+    public Integer getId() {
+        return id;
+    }
+
+    public String getNombre() {
+        return nombre;
+    }
+
+    public String getFuncion() {
+        return funcion;
+    }
+
+    public String getTelefono() {
+        return telefono;
+    }
+
+    public String getCorreo() {
+        return correo;
+    }
+
+    public Integer getHotelId() {
+        return hotelId;
+    }
 
     // Setters
-    public void setId(long id) { this.id = id; }
-    public void setNombre(String nombre) { this.nombre = nombre; }
-    public void setFuncion(String funcion) { this.funcion = funcion; }
-    public void setTelefono(String telefono) { this.telefono = telefono; }
-    public void setCorreo(String correo) { this.correo = correo; }
-    public void setHotelId(long hotelId) { this.hotelId = hotelId; }
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public void setNombre(String nombre) {
+        this.nombre = nombre;
+    }
+
+    public void setFuncion(String funcion) {
+        this.funcion = funcion;
+    }
+
+    public void setTelefono(String telefono) {
+        this.telefono = telefono;
+    }
+
+    public void setCorreo(String correo) {
+        this.correo = correo;
+    }
+
+    public void setHotelId(Integer hotelId) {
+        this.hotelId = hotelId;
+    }
 }
