+++ src/main/java/com/isst/mystay/repository/EmpleadoRepository.java
@@ -4,9 +4,8 @@ import com.isst.mystay.model.Empleado;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface EmpleadoRepository extends JpaRepository<Empleado, Long> {
-    List<Empleado> findByHotelId(long hotelId);
-
+public interface EmpleadoRepository extends JpaRepository<Empleado, Integer> {
+    List<Empleado> findByHotelId(Integer hotelId);
 
     Empleado findByFuncion(String funcion);
 }
