+++ src/main/java/com/isst/mystay/repository/HabitacionRepository.java
@@ -4,6 +4,6 @@ import com.isst.mystay.model.Habitacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface HabitacionRepository extends JpaRepository<Habitacion, Long> {
-    List<Habitacion> findByHotelId(long hotelId);
+public interface HabitacionRepository extends JpaRepository<Habitacion, Integer> {
+    List<Habitacion> findByHotelId(Integer hotelId);
 }
