+++ src/main/java/com/isst/mystay/repository/HotelRepository.java
@@ -3,5 +3,5 @@ package com.isst.mystay.repository;
 import com.isst.mystay.model.Hotel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HotelRepository extends JpaRepository<Hotel, Long> {
+public interface HotelRepository extends JpaRepository<Hotel, Integer> {
 }
