+++ src/main/java/com/isst/mystay/repository/RecursoRepository.java
@@ -4,16 +4,10 @@ import com.isst.mystay.model.Recurso;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface RecursoRepository extends JpaRepository<Recurso, Long> {
+public interface RecursoRepository extends JpaRepository<Recurso, Integer> {
     // Encuentra recursos por el nombre
     List<Recurso> findByNombre(String nombre);
 
     // Encuentra recursos asociados a un hotel específico por hotelId
-    List<Recurso> findByHotelId(Long hotelId);
+    List<Recurso> findByHotelId(Integer hotelId);
 }
-
-
-
-
-
-
